clean up azure resource generation
