+++ internal/cloud/azure/model.go
@@ -1,0 +1,22 @@
+package azure
+
+type ResourceStructure struct {
+	// Resource type name
+	ResourceTypeName string `json:"name"`
+	// Resource prefix as defined in the Azure Cloud Adoption Framework
+	CafPrefix string `json:"slug,omitempty"`
+	// MaxLength attribute define the maximum length of the name
+	MinLength int `json:"min_length"`
+	// MaxLength attribute define the maximum length of the name
+	MaxLength int `json:"max_length"`
+	// enforce lowercase
+	LowerCase bool `json:"lowercase,omitempty"`
+	// Regular expression to apply to the resource type
+	RegEx string `json:"regex,omitempty"`
+	// the Regular expression to validate the generated string
+	ValidationRegExp string `json:"validation_regex,omitempty"`
+	// can the resource include dashes
+	Dashes bool `json:"dashes"`
+	// The scope of this name where it needs to be unique
+	Scope string `json:"scope,omitempty"`
+}
