+++ tools/azure/templates/model.tmpl
@@ -6,38 +6,9 @@
 
 package azure
 
-type ResourceStructure struct {
-	// Resource type name
-	ResourceTypeName string `json:"name"`
-	// Resource prefix as defined in the Azure Cloud Adoption Framework
-	CafPrefix string `json:"slug,omitempty"`
-	// MaxLength attribute define the maximum length of the name
-	MinLength int `json:"min_length"`
-	// MaxLength attribute define the maximum length of the name
-	MaxLength int `json:"max_length"`
-	// enforce lowercase
-	LowerCase bool `json:"lowercase,omitempty"`
-	// Regular expression to apply to the resource type
-	RegEx string `json:"regex,omitempty"`
-	// the Regular expression to validate the generated string
-	ValidationRegExp string `json:"validation_regex,omitempty"`
-	// can the resource include dashes
-	Dashes bool `json:"dashes"`
-	// The scope of this name where it needs to be unique
-	Scope string `json:"scope,omitempty"`
-}
-
 // ResourceDefinitions are a map of difinitions for the resources supported
 var ResourceDefinitions = map[string]ResourceStructure{
     {{- range .ResourceStructures }}
     "{{.ResourceTypeName}}": {"{{.ResourceTypeName}}", "{{.CafPrefix}}", {{.MinLength}}, {{.MaxLength}},  {{.LowerCase}}, {{.RegEx}}, {{.ValidationRegExp}}, {{.Dashes}}, "{{.Scope}}" },
     {{- end}}
 }
-
-// ResourceMaps are a map from the slug to the resource definition
-var ResourceMaps = map[string]string {
-    {{- range $key, $value := .SlugMap}}
-        "{{$key}}": "{{$value}}",
-    {{- end}}
-}
-
